requirements.txt
matplotlib==1.4.3
nose==1.3.7
numpy==1.9.2
pyparsing==2.0.3
python-dateutil==2.4.2
python-magic==0.4.6
pytz==2015.4
six==1.9.0
